+++ package.json
@@ -58,7 +58,6 @@
     "webpack-hot-middleware": "^2.6.0"
   },
   "dependencies": {
-    "history": "^2.0.0",
     "material-ui": "^0.14.4",
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
