+++ src/index.js
@@ -5,13 +5,17 @@ import { Provider } from 'react-redux';
 import { createStore, compose, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import createBrowserHistory from 'history/lib/createBrowserHistory';
+import { browserHistory } from 'react-router'
 import contactsTable from './redux/reducer'
-//const store = createStore(contactsTable);
+import routes from './routes';
 
-const finalCreateStore = compose(applyMiddleware(thunkMiddleware))(createStore)
-const store = finalCreateStore(contactsTable)
+const finalCreateStore = compose(applyMiddleware(thunkMiddleware))(createStore);
+const store = finalCreateStore(contactsTable);
+
 render(
     <Provider store={store}>
-    <AppContainer />
+    <div>
+        {routes(store, browserHistory)}
+    </div>
     </Provider>
     , document.getElementById('root'));
